requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiogrpc==1.8
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astunparse==1.6.3
bcrypt==4.1.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click @ file:///private/tmp/python-click-20231211-45777-jkrufz/click-8.1.7
contourpy==1.2.0
control==0.9.4
cycler==0.12.1
dnspython==2.6.1
efficientnet-pytorch==0.7.1
email-validator==2.1.0.post1
fastapi==0.68.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gpxpy==1.6.2
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
haversine==2.8.1
huggingface-hub==0.20.3
idna==3.6
imageio==2.33.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
mavsdk==0.12.0
ml-dtypes==0.2.0
mpmath==1.3.0
munch==4.0.0
mysql-connector-python==8.2.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
passlib==1.7.4
pillow==10.2.0
pretrainedmodels==0.7.4
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML @ file:///private/tmp/pyyaml-20231018-5610-1rdh755/PyYAML-6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
segmentation-models-pytorch==0.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.14.2
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==0.9.12
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.15.0
Werkzeug==3.0.1
wrapt==1.14.1
